app/requirements.txt
# requirements.txt - Dipendenze aggiornate e sicure

# Framework web
Flask>=3.0.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
Jinja2>=3.1.2,<4.0.0
MarkupSafe>=2.1.3,<3.0.0
itsdangerous>=2.1.2,<3.0.0
click>=8.1.7,<9.0.0
blinker>=1.7.0,<2.0.0
watchdog>=3.0.0,<4.0.0

# Schedulazione
APScheduler>=3.10.4,<4.0.0
pytz>=2024.1
tzlocal>=5.2,<6.0.0

# Elaborazione dati
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.2,<4.0.0
xlrd>=2.0.1,<3.0.0

# Networking e HTTP
requests>=2.31.0,<3.0.0

# Utilità date e configurazione
python-dateutil>=2.8.2,<3.0.0
python-dotenv>=1.0.0,<2.0.0
six>=1.16.0,<2.0.0

# Logging strutturato
structlog>=24.1.0,<25.0.0

# Sicurezza aggiuntiva
cryptography>=42.0.0,<43.0.0

# Per info sistema e monitoring
psutil>=5.9.0,<6.0.0

# Development e testing (opzionale)
# pytest>=8.0.0,<9.0.0
# pytest-cov>=4.0.0,<5.0.0
# black>=24.0.0,<25.0.0
# flake8>=7.0.0,<8.0.0
# mypy>=1.8.0,<2.0.0

# Server FTP per testing (opzionale)
# pyftpdlib>=1.5.7,<2.0.0
